Use fake pom when adding Nexus deps in local mvnrepo
Latest version of the Maven install plugin will use the POM embedded inside the JAR file when installing the file in the local mvnrepo. Unfortunately, these extracted artifacts are Nexus PRO versions and make reference to a parent pom which is not publicly available. This will fail our build…

The solution here is to tell the mvn install:install-file plugin to generate a fake pom instead of using the one embedded in the jar.